test: move tests into UNit into the specific folder
