+++ tests/Popolazione/Actions/NomadelfiaActionTest.php
@@ -1,0 +1,59 @@
+<?php
+
+namespace Tests\Unit;
+
+use Carbon\Carbon;
+use Domain\Nomadelfia\PopolazioneNomadelfia\Actions\SaveEntrataInNomadelfiaAction;
+use Domain\Nomadelfia\PopolazioneNomadelfia\Actions\EntrataMinorenneAccoltoAction;
+use Domain\Nomadelfia\PopolazioneNomadelfia\Actions\EntrataMinorenneConFamigliaAction;
+use Domain\Nomadelfia\Famiglia\Models\Famiglia;
+use Domain\Nomadelfia\GruppoFamiliare\Models\GruppoFamiliare;
+use Domain\Nomadelfia\Persona\Models\Persona;
+use Tests\TestCase;
+
+it("entrata_minorenne_con_famiglia", function () {
+    $data_entrata = Carbon::now()->toDatestring();
+    $persona = Persona::factory()->minorenne()->femmina()->create();
+    $famiglia = Famiglia::factory()->create();
+
+    $gruppo = GruppoFamiliare::first();
+
+    $capoFam = Persona::factory()->maggiorenne()->maschio()->create();
+    $capoFam->gruppifamiliari()->attach($gruppo->id, ['stato' => '1', 'data_entrata_gruppo' => $data_entrata]);
+    $famiglia->assegnaCapoFamiglia($capoFam, $data_entrata);
+
+    $action = new EntrataMinorenneConFamigliaAction(new SaveEntrataInNomadelfiaAction());
+    $action->execute($persona, $data_entrata, Famiglia::findOrFail($famiglia->id));
+
+    $this->assertTrue($persona->isPersonaInterna());
+    $this->assertEquals($persona->getDataEntrataNomadelfia(), $data_entrata);
+//        $this->assertEquals($persona->posizioneAttuale()->id, $figlio->id);
+    $this->assertEquals($persona->posizioneAttuale()->pivot->data_inizio, $data_entrata);
+//        $this->assertEquals($persona->statoAttuale()->id, $nubile->id);
+//        $this->assertEquals($persona->statoAttuale()->stato, $nubile->stato);
+    $this->assertEquals($persona->statoAttuale()->pivot->data_inizio, $persona->data_nascita);
+    $this->assertEquals($persona->gruppofamiliareAttuale()->id, $gruppo->id);
+    $this->assertEquals($persona->gruppofamiliareAttuale()->pivot->data_entrata_gruppo, $data_entrata);
+    $this->assertNotNull($persona->famigliaAttuale());
+    $this->assertEquals($persona->famigliaAttuale()->pivot->data_entrata, $persona->data_nascita);
+
+});
+
+it("entrata_minorenne_accolto", function () {
+
+    $data_entrata = Carbon::now()->toDatestring();
+    $persona = Persona::factory()->minorenne()->femmina()->create();
+    $famiglia = Famiglia::factory()->create();
+
+    $gruppo = GruppoFamiliare::first();
+
+    $capoFam = Persona::factory()->maggiorenne()->maschio()->create();
+    $capoFam->gruppifamiliari()->attach($gruppo->id, ['stato' => '1', 'data_entrata_gruppo' => $data_entrata]);
+    $famiglia->assegnaCapoFamiglia($capoFam, $data_entrata);
+
+    $action = new EntrataMinorenneAccoltoAction(new SaveEntrataInNomadelfiaAction());
+    $action->execute($persona, $data_entrata, Famiglia::findOrFail($famiglia->id));
+
+    $this->assertTrue($persona->isPersonaInterna());
+
+});
